requirements.txt
aiogram
langchain
langchain-google-genai
langchain-chroma
langchain-community
langchain-core
python-dotenv